requirements.txt
Django==4.2.15
python-dotenv==1.0.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
boto3==1.28.62
django-storages==1.14.2